requirements.txt
certifi==2023.7.22
charset-normalizer==2.0.12
clarifai-grpc==7.0.0
dnspython==2.4.2
emoji==2.7.0
future==0.18.3
googleapis-common-protos==1.56.3
grpcio==1.48.2
idna==3.4
protobuf==3.19.6
pymongo==4.5.0
python-telegram-bot==10.1.0
requests==2.27.1
six==1.16.0
urllib3==1.26.16
